Model selection with CONGAS.

Selects the best number of clusters k to fit a dataset. Options allow to select the type of model available in the package, parameters for the Pyro optimizer and a number of different configurations to learn Bayesian posteriors or MAP estimates.

best_cluster(
  X,
  model,
  clusters,
  optim = "ClippedAdam",
  elbo = "TraceEnum_ELBO",
  inf_type = "SVI",
  steps = 300,
  lr = 0.05,
  param_list = list(),
  MAP = TRUE,
  seed = 3,
  mixture = NULL,
  method = "AIC",
  normalize_by_segs = FALSE
)

Arguments

X

rcongas object generated by the constructor init

model

Any of the model keywords available via list_models().

clusters

A vector of values for k.

optim

a pyro.optim object, in string format

elbo

a pyro.inference object, in string format

inf_type

currently just "SVI" is supported

steps

number of step of the gradient-based VI algorithm

lr

learning rate of the gradient-based VI algorithm

param_list

named list of parameters for the specific model

MAP

perform MAP or full bayesian inference (MAP is suggested as it converges better)

seed

seed for torch random number generator

mixture

initial mixture weights, if NULL then they are equal to 1/K

method

informtion criteria to score the best solution (currently supports "AIC", "BIC" and "ICL)

Value

A congas (described in [init]) object with following elements: - *inference* : - *models* with all the runs performed - *model_selection* with information about the IC and run statistics

Details

While this function has a lot of personalizable arguments, the default ones are usually good for most of the applications.