Example tibbles with input data in the format for (R)Congas+

Example tibbles with input data in the format for (R)Congas+

data(example_input)

Format

Example tibbles with input data.

Examples

data(example_input)
print(example_input)

#> $x_rna
#> # A tibble: 438,133 x 6
#>    gene     cell       value chr        from        to
#>    <chr>    <chr>      <int> <chr>     <int>     <int>
#>  1 DPM1     SRR5578183     1 chr20  50934867  50958555
#>  2 SCYL3    SRR5578183    10 chr1  169849631 169894267
#>  3 C1orf112 SRR5578183    26 chr1  169662007 169854080
#>  4 CFH      SRR5578183     6 chr1  196652043 196747504
#>  5 STPG1    SRR5578183    73 chr1   24356999  24416934
#>  6 NIPAL3   SRR5578183     5 chr1   24415802  24472976
#>  7 CFTR     SRR5578183     1 chr7  117287120 117715971
#>  8 CYP51A1  SRR5578183     3 chr7   92112153  92134803
#>  9 KRIT1    SRR5578183     3 chr7   92198969  92246166
#> 10 RAD52    SRR5578183     1 chr12    911736    990053
#> # … with 438,123 more rows
#> 
#> $x_atac
#> # A tibble: 27,303 x 6
#>    gene     cell             value chr        from        to
#>    <chr>    <chr>            <int> <chr>     <int>     <int>
#>  1 MRE11    SRR5578299 dfwfd    44 chr11  94415570  94493885
#>  2 HK1      SRR5578219 dfwfd    35 chr10  69269984  69401884
#>  3 MCPH1    SRR5578235 dfwfd    38 chr8    6406596   6648508
#>  4 ABCC6    SRR5578226 dfwfd    89 chr16  16148928  16223522
#>  5 ZNF213   SRR5578231 dfwfd    11 chr16   3129777   3142804
#>  6 CD84     SRR5578229 dfwfd     2 chr1  160541095 160579516
#>  7 KCTD10   SRR5578260 dfwfd    14 chr12 109448655 109477359
#>  8 RPS27AP1 SRR5578187 dfwfd    79 chr17  10258762  10259232
#>  9 UQCC3    SRR5578238 dfwfd    52 chr11  62670273  62673686
#> 10 KLHL21   SRR5578238 dfwfd    22 chr1    6590724   6614607
#> # … with 27,293 more rows
#> 
#> $x_segmentation
#> # A tibble: 22 x 4
#>    chr    from        to copies
#>    <chr> <int>     <int>  <int>
#>  1 chr1      0 249250621      2
#>  2 chr2      0 243199373      2
#>  3 chr3      0 198022430      2
#>  4 chr4      0 191154276      2
#>  5 chr5      0 180915260      2
#>  6 chr6      0 171115067      2
#>  7 chr7      0 159138663      2
#>  8 chr8      0 146364022      2
#>  9 chr9      0 141213431      2
#> 10 chr10     0 135534747      2
#> # … with 12 more rows
#> 
#> $x_normalisation_factors
#> # A tibble: 95 x 3
#>    cell       normalisation_factor modality
#>    <chr>                     <dbl> <chr>   
#>  1 SRR5578183                    1 RNA     
#>  2 SRR5578187                    1 RNA     
#>  3 SRR5578188                    1 RNA     
#>  4 SRR5578189                    1 RNA     
#>  5 SRR5578190                    1 RNA     
#>  6 SRR5578193                    1 RNA     
#>  7 SRR5578198                    1 RNA     
#>  8 SRR5578199                    1 RNA     
#>  9 SRR5578200                    1 RNA     
#> 10 SRR5578201                    1 RNA     
#> # … with 85 more rows
#>