Filter observed counts by quantile.

Filter observed counts by quantile.

filter_counts_by_quantile(x, upper_quantile = 0.98)

Arguments

x
An input RNA/ATAC dataset where entries are indexeable by genomic coordinate: "chr", "from" and "to".

upper_quantile
The maximum quantile to determine cuts. If a value is above the quantile it is removed

Value

The input data with removed entries.

Examples

data('example_input')
filter_counts_by_quantile(example_input$rna, upper_quantile = .98)

#> ── Upper quantile 0.98 

#> ℹ n = 5961 entries to remove

#> 
#> # A tibble: 5,961 x 8
#>    gene    chr       from       to cell                        value q_max del  
#>    <chr>   <chr>    <int>    <int> <chr>                       <int> <dbl> <lgl>
#>  1 ACAP3   chr1    1.29e6   1.31e6 bcc.su008.pre.tumor_AAGGCA…     2  1.88 TRUE 
#>  2 ZBTB40  chr1    2.24e7   2.25e7 bcc.su008.pre.tumor_AAGGCA…     2  1.8  TRUE 
#>  3 SRRM1   chr1    2.46e7   2.47e7 bcc.su008.pre.tumor_AAGGCA…    12 11.8  TRUE 
#>  4 TXNDC12 chr1    5.20e7   5.21e7 bcc.su008.pre.tumor_AAGGCA…     6  5.42 TRUE 
#>  5 ALG6    chr1    6.34e7   6.34e7 bcc.su008.pre.tumor_AAGGCA…     3  2.66 TRUE 
#>  6 SSX2IP  chr1    8.46e7   8.47e7 bcc.su008.pre.tumor_AAGGCA…     3  2.78 TRUE 
#>  7 MTF2    chr1    9.31e7   9.31e7 bcc.su008.pre.tumor_AAGGCA…     4  3.34 TRUE 
#>  8 GPSM2   chr1    1.09e8   1.09e8 bcc.su008.pre.tumor_AAGGCA…     8  6.5  TRUE 
#>  9 CELSR2  chr1    1.09e8   1.09e8 bcc.su008.pre.tumor_AAGGCA…     4  3.32 TRUE 
#> 10 RBM15   chr1    1.10e8   1.10e8 bcc.su008.pre.tumor_AAGGCA…     5  4.16 TRUE 
#> # … with 5,951 more rows

#> # A tibble: 195,498 x 6
#>    gene     chr      from      to cell                                 value
#>    <chr>    <chr>   <int>   <int> <chr>                                <int>
#>  1 NOC2L    chr1   944203  959309 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     2
#>  2 AGRN     chr1  1020120 1056118 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#>  3 SDF4     chr1  1216909 1232067 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#>  4 CPTP     chr1  1324756 1328896 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#>  5 AURKAIP1 chr1  1373730 1375495 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#>  6 CCNL2    chr1  1385711 1399335 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     2
#>  7 MRPL20   chr1  1401909 1407293 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#>  8 CDK11B   chr1  1635225 1659012 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#>  9 CDK11A   chr1  1702379 1724357 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#> 10 WRAP73   chr1  3630767 3652761 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     2
#> # … with 195,488 more rows

filter_counts_by_quantile(example_input$atac, upper_quantile = .98)

#> 

#> ── Upper quantile 0.98 

#> ℹ n = 60036 entries to remove

#> 
#> # A tibble: 60,036 x 7
#>    cell               value chr      from      to q_max del  
#>    <chr>              <int> <chr>   <int>   <int> <dbl> <lgl>
#>  1 SU008_Tumor_Pre_45     2 chr1   871996  872496  1.96 TRUE 
#>  2 SU008_Tumor_Pre_45     4 chr1   937126  937626  3.88 TRUE 
#>  3 SU008_Tumor_Pre_45     2 chr1  1050579 1051079  1.96 TRUE 
#>  4 SU008_Tumor_Pre_45     4 chr1  1138211 1138711  3.92 TRUE 
#>  5 SU008_Tumor_Pre_45     4 chr1  1176170 1176670  3.96 TRUE 
#>  6 SU008_Tumor_Pre_45     5 chr1  1186185 1186685  4.94 TRUE 
#>  7 SU008_Tumor_Pre_45     6 chr1  1238592 1239092  5.86 TRUE 
#>  8 SU008_Tumor_Pre_45     6 chr1  1239980 1240480  5.44 TRUE 
#>  9 SU008_Tumor_Pre_45     4 chr1  1241917 1242417  3.96 TRUE 
#> 10 SU008_Tumor_Pre_45     3 chr1  1280686 1281186  2.98 TRUE 
#> # … with 60,026 more rows

#> # A tibble: 521,771 x 5
#>    cell               value chr     from     to
#>    <chr>              <int> <chr>  <int>  <int>
#>  1 SU008_Tumor_Pre_45     1 chr1  127538 128038
#>  2 SU008_Tumor_Pre_45     2 chr1  540701 541201
#>  3 SU008_Tumor_Pre_45     4 chr1  762643 763143
#>  4 SU008_Tumor_Pre_45     3 chr1  859974 860474
#>  5 SU008_Tumor_Pre_45     1 chr1  866643 867143
#>  6 SU008_Tumor_Pre_45     2 chr1  876975 877475
#>  7 SU008_Tumor_Pre_45     3 chr1  894505 895005
#>  8 SU008_Tumor_Pre_45     2 chr1  895705 896205
#>  9 SU008_Tumor_Pre_45     1 chr1  898583 899083
#> 10 SU008_Tumor_Pre_45     3 chr1  901529 902029
#> # … with 521,761 more rows