Filter observed values by quantile.

Filter observed values by quantile.

filter_values_by_quantile(x, upper_quantile = 0.98)

Arguments

x
An input RNA/ATAC dataset where entries are indexeable by genomic coordinate: "chr", "from" and "to".

upper_quantile
The maximum quantile to determine cuts. If a value is above the quantile it is removed

Value

The input data with removed entries.

Examples

data('example_input')
filter_values_by_quantile(example_input$x_rna, upper_quantile = .98)

#> 

#> ── Upper quantile 0.98 

#> ℹ n = 16242 entries to remove

#> 
#> # A tibble: 16,242 x 8
#>    gene     cell       value chr        from        to   q_max del  
#>    <chr>    <chr>      <int> <chr>     <int>     <int>   <dbl> <lgl>
#>  1 M6PR     SRR5578183  7685 chr12   8940361   8949761  6741.  TRUE 
#>  2 KDM1A    SRR5578183 18970 chr1   23019443  23083689 11987.  TRUE 
#>  3 POLR2J   SRR5578183  3231 chr7  102473118 102478907  2991   TRUE 
#>  4 GCFC2    SRR5578183  3322 chr2   75652000  75710985  2989.  TRUE 
#>  5 ASB4     SRR5578183  2635 chr7   95478444  95540233  1496.  TRUE 
#>  6 OSBPL7   SRR5578183  6094 chr17  47807372  47821834  5679.  TRUE 
#>  7 CACNA2D2 SRR5578183    72 chr3   50362799  50504244    69.8 TRUE 
#>  8 SPAG9    SRR5578183 10838 chr17  50962174  51120868 10649.  TRUE 
#>  9 PKD1     SRR5578183  3499 chr16   2088708   2135898  2611.  TRUE 
#> 10 GIPR     SRR5578183   347 chr19  45668221  45683722   250.  TRUE 
#> # … with 16,232 more rows

#> Error: 

filter_values_by_quantile(example_input$x_atac, upper_quantile = .98)

#> 

#> ── Upper quantile 0.98 

#> ℹ n = 7282 entries to remove

#> 
#> # A tibble: 7,282 x 8
#>    gene   cell             value chr        from        to q_max del  
#>    <chr>  <chr>            <int> <chr>     <int>     <int> <dbl> <lgl>
#>  1 ABCC6  SRR5578226 dfwfd    89 chr16  16148928  16223522  87.1 TRUE 
#>  2 MTF1   SRR5578299 dfwfd    94 chr1   37809574  37859592  92.3 TRUE 
#>  3 RASA4  SRR5578222 dfwfd    84 chr7  102573807 102616756  81.4 TRUE 
#>  4 DMAP1  SRR5578228 dfwfd    80 chr1   44213455  44220681  79.2 TRUE 
#>  5 GRB2   SRR5578260 dfwfd    46 chr17  75318076  75405709  44.8 TRUE 
#>  6 GMPR   SRR5578286 dfwfd    64 chr6   16238587  16295549  63.2 TRUE 
#>  7 FLT3LG SRR5578271 dfwfd    68 chr19  49474207  49486231  67.4 TRUE 
#>  8 CCDC59 SRR5578249 dfwfd    80 chr12  82223681  82358805  80.0 TRUE 
#>  9 TXLNA  SRR5578218 dfwfd    69 chr1   32179675  32198285  67.0 TRUE 
#> 10 THAP7  SRR5578246 dfwfd   100 chr22  20999104  21002196  98.6 TRUE 
#> # … with 7,272 more rows

#> Error: