Extract the clusters ploidy profiles.

Extract the clone ploidy profiles from a fit object. It can subset by chromosomes, cluster and select what to highlight based on a parameter (alpha)

get_clusters_ploidy(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  clusters = NULL,
  offset_amplitude = TRUE,
  alpha = 0.05
)

Arguments

x
Input object with clusters.

chromosomes
Chromosome id to subset.

clusters
Cluster id to subset.

offset_amplitude
If TRUE, normalise CNA values for comparisons (z-score alike)

alpha
The parameter to select what to highlight.

Value

Examples

x = Rcongas::congas_example

# Default view
x %>% get_clusters_ploidy()
#> # A tibble: 140 × 9
#>    chr        from        to cluster       CN segment_mean offset_CN segment_id 
#>    <chr>     <dbl>     <dbl> <chr>      <dbl>        <dbl>     <dbl> <chr>      
#>  1 chr1          1  12600000 c1       0.0436          1.20   0.0436  chr1:1:126…
#>  2 chr1   12600001  60150000 c1       0.0165          1.58   0.0165  chr1:12600…
#>  3 chr1   61200001  77100000 c1      -0.0427          1.32  -0.0427  chr1:61200…
#>  4 chr1   78750001 156000000 c1       0.0184          1.59   0.0184  chr1:78750…
#>  5 chr1  156000001 249300000 c1       0.00960         1.51   0.00960 chr1:15600…
#>  6 chr2          1  26250000 c1       0.0797          1.62   0.0797  chr2:1:262…
#>  7 chr2   26250001  38100000 c1       0.0139          1.31   0.0139  chr2:26250…
#>  8 chr2   64050001  87150000 c1      -0.00652         1.49  -0.00652 chr2:64050…
#>  9 chr2   88350001 159900000 c1      -0.0418          1.52  -0.0418  chr2:88350…
#> 10 chr2  159900001 176850000 c1       0.0247          1.24   0.0247  chr2:15990…
#> # … with 130 more rows, and 1 more variable: highlight <lgl>

# Subset by chromosome
x %>% get_clusters_ploidy(chromosomes = 'chr1')
#> # A tibble: 10 × 9
#>    chr        from        to cluster       CN segment_mean offset_CN segment_id 
#>    <chr>     <dbl>     <dbl> <chr>      <dbl>        <dbl>     <dbl> <chr>      
#>  1 chr1          1  12600000 c1       0.0436          1.20   0.0436  chr1:1:126…
#>  2 chr1   12600001  60150000 c1       0.0165          1.58   0.0165  chr1:12600…
#>  3 chr1   61200001  77100000 c1      -0.0427          1.32  -0.0427  chr1:61200…
#>  4 chr1   78750001 156000000 c1       0.0184          1.59   0.0184  chr1:78750…
#>  5 chr1  156000001 249300000 c1       0.00960         1.51   0.00960 chr1:15600…
#>  6 chr1          1  12600000 c2      -0.0436          1.20  -0.0436  chr1:1:126…
#>  7 chr1   12600001  60150000 c2      -0.0165          1.58  -0.0165  chr1:12600…
#>  8 chr1   61200001  77100000 c2       0.0427          1.32   0.0427  chr1:61200…
#>  9 chr1   78750001 156000000 c2      -0.0184          1.59  -0.0184  chr1:78750…
#> 10 chr1  156000001 249300000 c2      -0.00960         1.51  -0.00960 chr1:15600…
#> # … with 1 more variable: highlight <lgl>

# Subset by cluster id and chromosome
x %>% get_clusters_ploidy(chromosomes = 'chr1', clusters = "c1")
#> # A tibble: 5 × 9
#>   chr        from        to cluster       CN segment_mean offset_CN segment_id  
#>   <chr>     <dbl>     <dbl> <chr>      <dbl>        <dbl>     <dbl> <chr>       
#> 1 chr1          1  12600000 c1       0.0436          1.20   0.0436  chr1:1:1260…
#> 2 chr1   12600001  60150000 c1       0.0165          1.58   0.0165  chr1:126000…
#> 3 chr1   61200001  77100000 c1      -0.0427          1.32  -0.0427  chr1:612000…
#> 4 chr1   78750001 156000000 c1       0.0184          1.59   0.0184  chr1:787500…
#> 5 chr1  156000001 249300000 c1       0.00960         1.51   0.00960 chr1:156000…
#> # … with 1 more variable: highlight <lgl>

# Change parameter to find what is most relevant
x %>% get_clusters_ploidy(alpha = 0.1)
#> # A tibble: 140 × 9
#>    chr        from        to cluster       CN segment_mean offset_CN segment_id 
#>    <chr>     <dbl>     <dbl> <chr>      <dbl>        <dbl>     <dbl> <chr>      
#>  1 chr1          1  12600000 c1       0.0436          1.20   0.0436  chr1:1:126…
#>  2 chr1   12600001  60150000 c1       0.0165          1.58   0.0165  chr1:12600…
#>  3 chr1   61200001  77100000 c1      -0.0427          1.32  -0.0427  chr1:61200…
#>  4 chr1   78750001 156000000 c1       0.0184          1.59   0.0184  chr1:78750…
#>  5 chr1  156000001 249300000 c1       0.00960         1.51   0.00960 chr1:15600…
#>  6 chr2          1  26250000 c1       0.0797          1.62   0.0797  chr2:1:262…
#>  7 chr2   26250001  38100000 c1       0.0139          1.31   0.0139  chr2:26250…
#>  8 chr2   64050001  87150000 c1      -0.00652         1.49  -0.00652 chr2:64050…
#>  9 chr2   88350001 159900000 c1      -0.0418          1.52  -0.0418  chr2:88350…
#> 10 chr2  159900001 176850000 c1       0.0247          1.24   0.0247  chr2:15990…
#> # … with 130 more rows, and 1 more variable: highlight <lgl>