Extract input data.

General data accessing getter function to return any of:

data,

segmentation,

normalisation factors.

The function uses the what parameter to return the appropriate type of information

Since the input formats are different, the outputs are also different based on the value of what, but they are always in tibble format.

Besides obvious columns, these information will also be available in the returned tibbles.

what = "data": modality and value_type, the latter reporting the likelihood code associated to the modality.

what = "segmentation":

ATAC_nonzerovals and ATAC_peaks, reporting the number of ATAC entries mapped to a segment, and the number of peaks these come from. Note that one non-zero entry is given by a cell that reported a ATAC count in a peak mapping within the segment.

RNA_nonzerovals and RNA_genes, reporting the number of RNA entries mapped to a segment, and the number of genes these come from. For RNA, one non-zero entry is given by a cell that reported an RNA count in a gene mapping within the segment.

what = "normalisation": just modality.

get_input(x, what = "data")

Arguments

x
An object of class rcongasplus.

what
Any of "data", "segmentation" or "normalisation".

Value

A tibble; its format depends on what. See the examples.

Examples

data(example_object)

# Extract the input data, after mapping to segments
get_input(example_object, what = 'data') 

#> # A tibble: 29,190 x 5
#>    segment_id            cell                          value modality value_type
#>    <chr>                 <chr>                         <int> <chr>    <chr>     
#>  1 chr1:57781319:108512… bcc.su008.pre.cd45_AAAGATGCA…   274 RNA      NB        
#>  2 chr1:57781319:108512… bcc.su008.pre.cd45_GCAATCATC…   227 RNA      NB        
#>  3 chr1:57781319:108512… bcc.su008.pre.cd45_GCAGCCATC…   209 RNA      NB        
#>  4 chr1:57781319:108512… bcc.su008.pre.cd45_TCCACACCA…   258 RNA      NB        
#>  5 chr1:57781319:108512… bcc.su008.pre.tcell_AGCATACC…   284 RNA      NB        
#>  6 chr1:57781319:108512… bcc.su008.pre.tcell_AGCTCCTA…    56 RNA      NB        
#>  7 chr1:57781319:108512… bcc.su008.pre.tcell_GGAAAGCG…   130 RNA      NB        
#>  8 chr1:57781319:108512… bcc.su008.pre.tcell_GGTGCGTC…   195 RNA      NB        
#>  9 chr1:57781319:108512… bcc.su008.pre.tumor_AAACCTGT…   251 RNA      NB        
#> 10 chr1:57781319:108512… bcc.su008.pre.tumor_AAAGCAAG…   256 RNA      NB        
#> # … with 29,180 more rows

# e.g., in this way you can get input ATAC values
get_input(example_object, what = 'data') %>% dplyr::filter(modality == 'ATAC')

#> # A tibble: 7,770 x 5
#>    segment_id              cell                value modality value_type
#>    <chr>                   <chr>               <int> <chr>    <chr>     
#>  1 chr1:57781319:108512982 SU008_Tumor_Pre_10   1961 ATAC     NB        
#>  2 chr1:57781319:108512982 SU008_Tumor_Pre_100   342 ATAC     NB        
#>  3 chr1:57781319:108512982 SU008_Tumor_Pre_104   602 ATAC     NB        
#>  4 chr1:57781319:108512982 SU008_Tumor_Pre_106   511 ATAC     NB        
#>  5 chr1:57781319:108512982 SU008_Tumor_Pre_110   335 ATAC     NB        
#>  6 chr1:57781319:108512982 SU008_Tumor_Pre_112   593 ATAC     NB        
#>  7 chr1:57781319:108512982 SU008_Tumor_Pre_114   794 ATAC     NB        
#>  8 chr1:57781319:108512982 SU008_Tumor_Pre_115   273 ATAC     NB        
#>  9 chr1:57781319:108512982 SU008_Tumor_Pre_116  1076 ATAC     NB        
#> 10 chr1:57781319:108512982 SU008_Tumor_Pre_118   910 ATAC     NB        
#> # … with 7,760 more rows

 
# Extract the input segmentation. 
get_input(example_object, what = 'segmentation')

#> # A tibble: 30 x 12
#>    chr       from       to copies segment_id         ATAC_nonzerovals ATAC_peaks
#>    <chr>    <int>    <int>  <int> <chr>                         <dbl>      <dbl>
#>  1 chr10   2.85e5   1.35e8      2 chr10:285200:1354…           176459      18251
#>  2 chr4    4.03e4   1.91e8      2 chr4:40332:190986…           153390      16652
#>  3 chr11   1.75e7   1.35e8      3 chr11:17515668:13…           180071      17098
#>  4 chr2    1.33e8   2.42e8      3 chr2:133074851:24…           122575      13415
#>  5 chr2    3.93e4   8.93e7      3 chr2:39340:893198…           130226      12956
#>  6 chr14   2.03e7   1.05e8      3 chr14:20323479:10…           125378      12371
#>  7 chr15   2.89e7   1.03e8      2 chr15:28947721:10…           132411      12616
#>  8 chr6    6.73e7   1.71e8      2 chr6:67340783:171…           105050      11560
#>  9 chr12   5.33e7   1.34e8      3 chr12:53343461:13…           120836      11634
#> 10 chr5    1.10e8   1.81e8      3 chr5:110286431:18…           105659      10740
#> # … with 20 more rows, and 5 more variables: RNA_nonzerovals <dbl>,
#> #   RNA_genes <dbl>, E <dbl>, L <int>, row_number() <= 30 <lgl>

 
# Extract the input normalisation factors
get_input(example_object, what = 'normalisation')

#> # A tibble: 973 x 3
#>    cell                                 normalisation_factor modality
#>    <chr>                                               <dbl> <chr>   
#>  1 bcc.su008.pre.cd45_AAAGATGCAAAGGAAG                 2.70  RNA     
#>  2 bcc.su008.pre.cd45_GCAATCATCAAACAAG                 1.89  RNA     
#>  3 bcc.su008.pre.cd45_GCAGCCATCTTTAGTC                 1.66  RNA     
#>  4 bcc.su008.pre.cd45_TCCACACCACATCTTT                 2.49  RNA     
#>  5 bcc.su008.pre.tcell_AGCATACCAGCATGAG                1.59  RNA     
#>  6 bcc.su008.pre.tcell_AGCTCCTAGGACAGAA                0.614 RNA     
#>  7 bcc.su008.pre.tcell_GGAAAGCGTATTAGCC                1.15  RNA     
#>  8 bcc.su008.pre.tcell_GGTGCGTCAGAGTGTG                1.74  RNA     
#>  9 bcc.su008.pre.tumor_AAACCTGTCCGAGCCA                1.17  RNA     
#> 10 bcc.su008.pre.tumor_AAAGCAAGTGTCTGAT                0.762 RNA     
#> # … with 963 more rows