Extract the segment ids (for indexing).

For the functions that require a segment identifier as input, for instance plot_segment_density, this function returns the list of segment ids used in the input object. If run with highlight = TRUE it will return only the segments that are highlighted in the plots; this requires clustering to be computed

get_segment_ids(x, highlight = FALSE, ...)

Arguments

x
Input object (with or without clustering).

highlight
If TRUE, returns return only the segments that are highlighted in the plots. Use ellipsis to specify parameters such as alpha to function get_clusters_ploidy.

...
Parameters forwarded to get_clusters_ploidy to get segment ids.

Value

A vector of strings.

Examples

x = Rcongas::congas_example

# All segments
get_segment_ids(x)
#>  [1] "chr1:1:12600000"           "chr1:12600001:60150000"   
#>  [3] "chr1:61200001:77100000"    "chr1:78750001:156000000"  
#>  [5] "chr1:156000001:249300000"  "chr2:1:26250000"          
#>  [7] "chr2:26250001:38100000"    "chr2:64050001:87150000"   
#>  [9] "chr2:88350001:159900000"   "chr2:159900001:176850000" 
#> [11] "chr2:176850001:195450000"  "chr2:201750001:243300000" 
#> [13] "chr3:1:13800000"           "chr3:34350001:86850000"   
#> [15] "chr3:89100001:114000000"   "chr3:114000001:134100000" 
#> [17] "chr3:134700001:152100000"  "chr3:152100001:180450000" 
#> [19] "chr3:180450001:198150000"  "chr4:1:70500000"          
#> [21] "chr4:70500001:115350000"   "chr4:135300001:151650000" 
#> [23] "chr4:151650001:191250000"  "chr5:18000001:54150000"   
#> [25] "chr5:54150001:127050000"   "chr5:129150001:150600000" 
#> [27] "chr5:166500001:181050000"  "chr6:1:14850000"          
#> [29] "chr6:14850001:32400000"    "chr6:32700001:76950000"   
#> [31] "chr6:93450001:114450000"   "chr6:145500001:156600000" 
#> [33] "chr7:4950001:132450000"    "chr7:135300001:156900000" 
#> [35] "chr8:1:67500000"           "chr8:100800001:146400000" 
#> [37] "chr9:1:78900000"           "chr9:85350001:141300000"  
#> [39] "chr10:1500001:76200000"    "chr10:76950001:118350000" 
#> [41] "chr10:118350001:135600000" "chr11:1:33750000"         
#> [43] "chr11:43500001:100200000"  "chr11:100200001:135150000"
#> [45] "chr12:1:14400000"          "chr12:16200001:46800000"  
#> [47] "chr12:46800001:68850000"   "chr12:68850001:133950000" 
#> [49] "chr13:21150001:38850000"   "chr13:38850001:83400000"  
#> [51] "chr14:1:93300000"          "chr14:93300001:105450000" 
#> [53] "chr15:1:58950000"          "chr15:67050001:102600000" 
#> [55] "chr16:1:3750000"           "chr16:8550001:15600000"   
#> [57] "chr16:16950001:90450000"   "chr17:1:40650000"         
#> [59] "chr17:40650001:47250000"   "chr17:47250001:79650000"  
#> [61] "chr17:79650001:81300000"   "chr18:7950001:20400000"   
#> [63] "chr18:32400001:55950000"   "chr18:55950001:78150000"  
#> [65] "chr19:1:59250000"          "chr20:1:13650000"         
#> [67] "chr20:26100001:42750000"   "chr20:50550001:63150000"  
#> [69] "chr21:1:38850000"          "chr22:1:51450000"         

# All segments that are highlighted with default value
get_segment_ids(x, highlight = TRUE)
#> [1] "chr14:93300001:105450000" "chr15:67050001:102600000"
#> [3] "chr16:1:3750000"          "chr18:32400001:55950000" 

# All segments that are highlighted with stricter value than default
get_segment_ids(x, highlight = TRUE, alpha = 0.01)
#> [1] "chr15:67050001:102600000" "chr16:1:3750000"         
#> [3] "chr18:32400001:55950000"