Wilcoxon test for difference in RNA counts among groups.

Performs a wilcox.test for the counts in each segments, split by pair of groups. A group can be a list of cluster ids; the test p-value can be passed as a parameter, and is adjusted for Family-Wise Error Rate (FWER) via Bonferroni correction.

get_segment_test_counts(x, group1, group2, cutoff_p = 0.01)

Arguments

x
Input object with clusters.

group1
Group of cluster ids (vs group2).

group2
Group of cluster ids (vs group1).

cutoff_p
P-value cutoff

Value

A tibble with segment coordinates, p-values and a sign status for the significance of the test.

Examples

x = Rcongas::congas_example

# Compare groups
get_segment_test_counts(x, "c1", "c2")
#> # A tibble: 70 × 5
#>    chr       from        to        p sign 
#>    <chr>    <int>     <int>    <dbl> <lgl>
#>  1 chr15 67050001 102600000 4.11e-59 TRUE 
#>  2 chr18 32400001  55950000 1.11e-57 TRUE 
#>  3 chr16        1   3750000 1.26e-57 TRUE 
#>  4 chr14 93300001 105450000 5.32e-50 TRUE 
#>  5 chr16 16950001  90450000 3.82e-49 TRUE 
#>  6 chr19        1  59250000 6.20e-49 TRUE 
#>  7 chr11 43500001 100200000 2.38e-48 TRUE 
#>  8 chr1  12600001  60150000 2.67e-47 TRUE 
#>  9 chr6  93450001 114450000 1.82e-45 TRUE 
#> 10 chr7   4950001 132450000 5.27e-45 TRUE 
#> # … with 60 more rows

# More stringent test
get_segment_test_counts(x, "c1", "c2", cutoff_p = 1e-4)
#> # A tibble: 70 × 5
#>    chr       from        to        p sign 
#>    <chr>    <int>     <int>    <dbl> <lgl>
#>  1 chr15 67050001 102600000 4.11e-59 TRUE 
#>  2 chr18 32400001  55950000 1.11e-57 TRUE 
#>  3 chr16        1   3750000 1.26e-57 TRUE 
#>  4 chr14 93300001 105450000 5.32e-50 TRUE 
#>  5 chr16 16950001  90450000 3.82e-49 TRUE 
#>  6 chr19        1  59250000 6.20e-49 TRUE 
#>  7 chr11 43500001 100200000 2.38e-48 TRUE 
#>  8 chr1  12600001  60150000 2.67e-47 TRUE 
#>  9 chr6  93450001 114450000 1.82e-45 TRUE 
#> 10 chr7   4950001 132450000 5.27e-45 TRUE 
#> # … with 60 more rows