Glioblastoma tumour and normal (SmartSeq RNA, z-score)

@description (R)CONGAS+ object for the glioblastoma dataset, first used in the original CONGAS paper. The paper providing these data is Patel, Anoop P., et al. "Single-cell RNA-seq highlights intratumoral heterogeneity in primary glioblastoma." Science 344.6190 (2014): 1396-1401.

data(patel_gbmtn)

Format

(R)CONGAS+ object.

Examples

data(patel_gbmtn)
print(patel_gbmtn)
#> ── [ (R)CONGAS+ ] Patel et al. GBM tumour/normal used in CONGAS (SmartSeq) ─────
#> 
#> ── CNA segments (reference: GRCh38) 
#> → Input 14 CNA segments, mean ploidy 2.
#> 
#> 	 | | |  | | | | | |  |  
#> 
#> 	 Ploidy:    0     1     2     3     4     5     *   
#> 
#> ── Modalities 
#> → RNA: 75 cells with 1506 mapped genes, 62416 non-zero values. Likelihood: Gaussian.
#> → ATAC: not available
#> ! Clusters: not available.
#> 
#> ──  LOG  ──
#> 
#> - 2021-04-02 10:32:31 Created input object.
#> [1] 0