Run a complete simulation

This function runs a n entre simulation, it first segments the genome, then it assigns copy number to each clone based on the given tree

run_simulation(
  cnv_df,
  ncells = 1000,
  props = c(0.8, 0.2),
  K = 2,
  theta_loc = 23000,
  theta_shape = 0.3,
  model_matrix = NULL,
  coeff = c(0),
  class_rand = c(2, 2),
  k_rand = 2,
  random_cov = FALSE,
  het_snps_for_gene = 1,
  prob_random = c(0.99, 0.1),
  random_class_dim = c(0.2, 0.1),
  perc_dropout = 0,
  baf_var = 20,
  gamma_shape = 0.45,
  gamma_rate = 0.05,
  monoallelic_rate = 0.8,
  haplotypes = trivial_hap(cnv_df),
  max_snps = 5,
  fast = FALSE,
  gamma_shape_fast = 6,
  gamma_rate_fast = 6,
  theta_shape_fast = 6,
  theta_rate_fast = 4
)

Arguments

ncells
integer, number of cells to simulate

props
proportions of each cluster, needs to have length = K

theta_shape
shape parameter for the Gamma modelling the library factor

theta_rate
rate parameter for the Gamma modelling the library factor

Value

a CNVSimulation object

Examples