Basal cell carcinoma (10x RNA and ATAC, counts-based)

(R)CONGAS+ object for the basal cell carcinoma dataset, used in the CONGAS+ paper. The paper providing these data is Satpathy, Ansuman T., et al. "Massively parallel single-cell chromatin landscapes of human immune cell development and intratumoral T cell exhaustion." Nature Biotechnology 37.8 (2019): 925-936.

data(satpathy_bcc)

Format

(R)CONGAS+ object.

Examples

data(satpathy_bcc)
print(satpathy_bcc)

#> ── [ (R)CONGAS+ ] SU008 TUMOR 30 segments (73.66% genome) ──────────────────────

#> 

#> ── CNA segments (reference: GRCh38) 

#> → Input 30 CNA segments, mean ploidy 3.2.

#> 
#> 	 | | |  | | | |  | | | |  | | |  | |  |   |  |  | 
#> 
#> 	 Ploidy:    0     1     2     3     4     5     *   

#> 

#> ── Modalities 

#> → RNA: 714 cells with 8613 mapped genes, 1401728 non-zero values. Likelihood: Negative Binomial.

#> → ATAC: 259 cells with 284316 mapped peaks, 3083691 non-zero values. Likelihood: Negative Binomial.

#> ! Clusters: not available.

#> 

#> ──  LOG  ──

#> 

#> - 2021-03-30 17:58:41 Created input object.
#> - 2021-03-30 17:58:43 Filtered outliers: [6|0.05|0.95]

#> [1] 0