Compute the statistics of a dataset.

The statistics of a dataset regard the input data, and the fits. These are accessed by using different values for the what parameter passed to this function (by default, data are used).

The result value is a named list with several information that are usefull to summarise the model. The names should be self explicatory.

For data the following information are reported:

it is reported the number of cells in each modality,

the number of non-zero ATAC/RNA input entries used to create the object

the number of RNA genes and ATAC peaks used

the number of segments

the number of modalities

the keyword for each modality ("RNA" or "ATAC")

the likelihood for each modality ("G" for Gaussian or "NB" for Negative Binomial)

For fits the following information are reported:

...

If fits are not available, NULL is returned.

stat(x, what = "data")

Arguments

x

An object of class rcongasplus.

what

The type of statistics that needs to be computed. It can be one of "data" or "fit". By deafult it is "data".

Value

A named list with the available statistics.

Examples

data(example_object)

# Compare the information here 
print(example_object)
#> Error in sanitize_zeroes(.): could not find function "sanitize_zeroes"

# ... to the one reported here for data
stat(example_object, what = 'data')
#> Error in sanitize_obj(.): could not find function "sanitize_obj"

# ... and here for fits
stat(example_object, what = 'fit')
#> Error in sanitize_obj(.): could not find function "sanitize_obj"