Example tibbles with input data in the format for (R)CONGAS+

Example tibbles with input data in the format for (R)CONGAS+

data(example_input)

Format

Example tibbles with input data.

Examples

data(example_input)
print(example_input)

#> $rna
#> # A tibble: 201,459 x 6
#>    gene     chr      from      to cell                                 value
#>    <chr>    <chr>   <int>   <int> <chr>                                <int>
#>  1 NOC2L    chr1   944203  959309 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     2
#>  2 AGRN     chr1  1020120 1056118 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#>  3 SDF4     chr1  1216909 1232067 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#>  4 ACAP3    chr1  1292390 1309609 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     2
#>  5 CPTP     chr1  1324756 1328896 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#>  6 AURKAIP1 chr1  1373730 1375495 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#>  7 CCNL2    chr1  1385711 1399335 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     2
#>  8 MRPL20   chr1  1401909 1407293 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#>  9 CDK11B   chr1  1635225 1659012 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#> 10 CDK11A   chr1  1702379 1724357 bcc.su008.pre.tumor_AAGGCAGTCACCGTAA     1
#> # … with 201,449 more rows
#> 
#> $atac
#> # A tibble: 581,807 x 5
#>    cell               value chr     from     to
#>    <chr>              <int> <chr>  <int>  <int>
#>  1 SU008_Tumor_Pre_45     1 chr1  127538 128038
#>  2 SU008_Tumor_Pre_45     2 chr1  540701 541201
#>  3 SU008_Tumor_Pre_45     4 chr1  762643 763143
#>  4 SU008_Tumor_Pre_45     3 chr1  859974 860474
#>  5 SU008_Tumor_Pre_45     1 chr1  866643 867143
#>  6 SU008_Tumor_Pre_45     2 chr1  871996 872496
#>  7 SU008_Tumor_Pre_45     2 chr1  876975 877475
#>  8 SU008_Tumor_Pre_45     3 chr1  894505 895005
#>  9 SU008_Tumor_Pre_45     2 chr1  895705 896205
#> 10 SU008_Tumor_Pre_45     1 chr1  898583 899083
#> # … with 581,797 more rows
#> 
#> $segments
#> # A tibble: 30 x 4
#>    chr        from        to copies
#>    <chr>     <int>     <int>  <int>
#>  1 chr10    285200 135490954      2
#>  2 chr4      40332 190986668      2
#>  3 chr11  17515668 134898011      3
#>  4 chr2  133074851 241531737      3
#>  5 chr2      39340  89319834      3
#>  6 chr14  20323479 105411153      3
#>  7 chr15  28947721 102517302      2
#>  8 chr6   67340783 171046023      2
#>  9 chr12  53343461 133801333      3
#> 10 chr5  110286431 180712485      3
#> # … with 20 more rows
#>