Example tibbles with input data in the format for (R)Congas+

Example tibbles with input data in the format for (R)Congas+

data(example_input)

Format

Example tibbles with input data.

Examples

data(example_input)
print(example_input)

#> $x_rna
#> # A tibble: 438,133 x 6
#>    gene     cell       count chr        from        to
#>    <chr>    <chr>      <int> <chr>     <int>     <int>
#>  1 DPM1     SRR5578183     1 chr20  50934867  50958555
#>  2 SCYL3    SRR5578183    10 chr1  169849631 169894267
#>  3 C1orf112 SRR5578183    26 chr1  169662007 169854080
#>  4 CFH      SRR5578183     6 chr1  196652043 196747504
#>  5 STPG1    SRR5578183    73 chr1   24356999  24416934
#>  6 NIPAL3   SRR5578183     5 chr1   24415802  24472976
#>  7 CFTR     SRR5578183     1 chr7  117287120 117715971
#>  8 CYP51A1  SRR5578183     3 chr7   92112153  92134803
#>  9 KRIT1    SRR5578183     3 chr7   92198969  92246166
#> 10 RAD52    SRR5578183     1 chr12    911736    990053
#> # … with 438,123 more rows
#> 
#> $x_atac
#> # A tibble: 27,019 x 6
#>    gene     cell             count chr        from        to
#>    <chr>    <chr>            <int> <chr>     <int>     <int>
#>  1 COX6A1   SRR5578193 dfwfd    48 chr12 120438090 120440737
#>  2 RRP1     SRR5578286 dfwfd    61 chr21  43789513  43805293
#>  3 RPS3AP47 SRR5578224 dfwfd    53 chr15  43115702  43116493
#>  4 FECH     SRR5578244 dfwfd    37 chr18  57544377  57586702
#>  5 SLC35F1  SRR5578221 dfwfd    98 chr6  117907264 118317676
#>  6 RBM28    SRR5578228 dfwfd    85 chr7  128297685 128343908
#>  7 C11orf54 SRR5578198 dfwfd    85 chr11  93741591  93764749
#>  8 AASDHPPT SRR5578244 dfwfd    30 chr11 106075501 106098699
#>  9 NPFFR1   SRR5578219 dfwfd    24 chr10  70247329  70284004
#> 10 HEATR1   SRR5578236 dfwfd    43 chr1  236549005 236604516
#> # … with 27,009 more rows
#> 
#> $x_segmentation
#> # A tibble: 22 x 4
#>    chr    from        to copies
#>    <chr> <int>     <int>  <int>
#>  1 chr1      0 249250621      2
#>  2 chr2      0 243199373      2
#>  3 chr3      0 198022430      2
#>  4 chr4      0 191154276      2
#>  5 chr5      0 180915260      2
#>  6 chr6      0 171115067      2
#>  7 chr7      0 159138663      2
#>  8 chr8      0 146364022      2
#>  9 chr9      0 141213431      2
#> 10 chr10     0 135534747      2
#> # … with 12 more rows
#> 
#> $x_normalisation_factors
#>                 cell normalisation_factor modality
#> 1         SRR5578183                    1      RNA
#> 2         SRR5578187                    1      RNA
#> 3         SRR5578188                    1      RNA
#> 4         SRR5578189                    1      RNA
#> 5         SRR5578190                    1      RNA
#> 6         SRR5578193                    1      RNA
#> 7         SRR5578198                    1      RNA
#> 8         SRR5578199                    1      RNA
#> 9         SRR5578200                    1      RNA
#> 10        SRR5578201                    1      RNA
#> 11        SRR5578203                    1      RNA
#> 12        SRR5578205                    1      RNA
#> 13        SRR5578206                    1      RNA
#> 14        SRR5578207                    1      RNA
#> 15        SRR5578209                    1      RNA
#> 16        SRR5578210                    1      RNA
#> 17        SRR5578214                    1      RNA
#> 18        SRR5578216                    1      RNA
#> 19        SRR5578217                    1      RNA
#> 20        SRR5578218                    1      RNA
#> 21        SRR5578219                    1      RNA
#> 22        SRR5578221                    1      RNA
#> 23        SRR5578222                    1      RNA
#> 24        SRR5578224                    1      RNA
#> 25        SRR5578226                    1      RNA
#> 26        SRR5578227                    1      RNA
#> 27        SRR5578228                    1      RNA
#> 28        SRR5578229                    1      RNA
#> 29        SRR5578230                    1      RNA
#> 30        SRR5578231                    1      RNA
#> 31        SRR5578233                    1      RNA
#> 32        SRR5578235                    1      RNA
#> 33        SRR5578236                    1      RNA
#> 34        SRR5578238                    1      RNA
#> 35        SRR5578239                    1      RNA
#> 36        SRR5578240                    1      RNA
#> 37        SRR5578243                    1      RNA
#> 38        SRR5578244                    1      RNA
#> 39        SRR5578246                    1      RNA
#> 40        SRR5578248                    1      RNA
#> 41        SRR5578249                    1      RNA
#> 42        SRR5578257                    1      RNA
#> 43        SRR5578258                    1      RNA
#> 44        SRR5578260                    1      RNA
#> 45        SRR5578271                    1      RNA
#> 46        SRR5578278                    1      RNA
#> 47        SRR5578280                    1      RNA
#> 48        SRR5578281                    1      RNA
#> 49        SRR5578282                    1      RNA
#> 50        SRR5578286                    1      RNA
#> 51        SRR5578292                    1      RNA
#> 52        SRR5578298                    1      RNA
#> 53        SRR5578299                    1      RNA
#> 54  SRR5578193 dfwfd                    1     ATAC
#> 55  SRR5578210 dfwfd                    1     ATAC
#> 56  SRR5578205 dfwfd                    1     ATAC
#> 57  SRR5578286 dfwfd                    1     ATAC
#> 58  SRR5578233 dfwfd                    1     ATAC
#> 59  SRR5578243 dfwfd                    1     ATAC
#> 60  SRR5578224 dfwfd                    1     ATAC
#> 61  SRR5578244 dfwfd                    1     ATAC
#> 62  SRR5578183 dfwfd                    1     ATAC
#> 63  SRR5578221 dfwfd                    1     ATAC
#> 64  SRR5578228 dfwfd                    1     ATAC
#> 65  SRR5578198 dfwfd                    1     ATAC
#> 66  SRR5578219 dfwfd                    1     ATAC
#> 67  SRR5578236 dfwfd                    1     ATAC
#> 68  SRR5578235 dfwfd                    1     ATAC
#> 69  SRR5578249 dfwfd                    1     ATAC
#> 70  SRR5578226 dfwfd                    1     ATAC
#> 71  SRR5578292 dfwfd                    1     ATAC
#> 72  SRR5578248 dfwfd                    1     ATAC
#> 73  SRR5578238 dfwfd                    1     ATAC
#> 74  SRR5578298 dfwfd                    1     ATAC
#> 75  SRR5578206 dfwfd                    1     ATAC
#> 76  SRR5578299 dfwfd                    1     ATAC
#> 77  SRR5578203 dfwfd                    1     ATAC
#> 78  SRR5578281 dfwfd                    1     ATAC
#> 79  SRR5578214 dfwfd                    1     ATAC
#> 80  SRR5578189 dfwfd                    1     ATAC
#> 81  SRR5578239 dfwfd                    1     ATAC
#> 82  SRR5578282 dfwfd                    1     ATAC
#> 83  SRR5578190 dfwfd                    1     ATAC
#> 84  SRR5578207 dfwfd                    1     ATAC
#> 85  SRR5578246 dfwfd                    1     ATAC
#> 86  SRR5578227 dfwfd                    1     ATAC
#> 87  SRR5578201 dfwfd                    1     ATAC
#> 88  SRR5578217 dfwfd                    1     ATAC
#> 89  SRR5578222 dfwfd                    1     ATAC
#> 90  SRR5578216 dfwfd                    1     ATAC
#> 91  SRR5578199 dfwfd                    1     ATAC
#> 92  SRR5578280 dfwfd                    1     ATAC
#> 93  SRR5578230 dfwfd                    1     ATAC
#> 94  SRR5578187 dfwfd                    1     ATAC
#> 95  SRR5578257 dfwfd                    1     ATAC
#> 96  SRR5578229 dfwfd                    1     ATAC
#> 97  SRR5578258 dfwfd                    1     ATAC
#> 98  SRR5578278 dfwfd                    1     ATAC
#> 99  SRR5578271 dfwfd                    1     ATAC
#> 100 SRR5578231 dfwfd                    1     ATAC
#> 101 SRR5578260 dfwfd                    1     ATAC
#> 102 SRR5578200 dfwfd                    1     ATAC
#> 103 SRR5578240 dfwfd                    1     ATAC
#> 104 SRR5578188 dfwfd                    1     ATAC
#> 105 SRR5578218 dfwfd                    1     ATAC
#> 106 SRR5578209 dfwfd                    1     ATAC
#>