Fit a list of CONGAS model

This function is general interface for fitting a congas model in R. The model briefly consist in a joint mixture model over two modalities, currently scATAC and scRNA-seq. For more information about the theoretical fundations of the approach refer to the vignette. This function performs modele selection over a specified number of clusters, using a specific information criterium (IC). ICs and results for all the runs are, however, reported in the object.

fit_congas(
  x,
  K,
  model_parameters,
  learning_rate = 0.05,
  latent_variables = "D",
  compile = FALSE,
  steps = 500,
  model_selection = "ICL",
  temperature = 10
)

Arguments

x
An rcongasplus object with the input dataset, constructed with Rcongas::init.

K
a vector of integers with the number of clusters we want to test

model_parameters
a list with model hyperparameters. As errors coming from wrong hyperparameters initialization are quite hard to troubleshoot is higly suggested to use Rcongas::auto_config_run() to generate a template and eventually modify it.

learning_rate
a learning rate for the Adam optimizer

latent_variables
specify the nature of the latent variable modelling the copy number profile. Currently only "D" (discrete) is available

compile
use JIT compiler for the Pyro ba

steps
number of steps of optimization

model_selection
information criteria to which perform the model selection (one of ICL, NLL, BIC, AIC)

Value

An object ot class rcongasplus with a slot bset_fit with the learned parameters for the selected model in tiblle format. A slot runs with all the runs performed ordered by the selectde IC and a slot model_selection with all the information to perform model selection.

Details

The functions assume a list of model hyperparameters. As the the model formulation isquite complex, and those hyperparameters are extremely difficult to set by hand we suggest the usage of the function Rcongas::auto_config_run()

Examples

library(Rcongas)
if (FALSE) {
K <-  1:4
hyperparams <- auto_config_run(example_object, 1:4)

fit <- fit_congas(example_object, K = 1:4,learning_rate = 0.05, model_parameters = hyperparams)
}