Extract the clusters ploidy profiles.

Extract the clone ploidy profiles from a fit object. It can subset by chromosomes, cluster and select what to highlight based on a parameter (alpha)

get_clusters_ploidy(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  clusters = NULL,
  offset_amplitude = TRUE,
  alpha = 0.05
)

Arguments

x
Input object with clusters.

chromosomes
Chromosome id to subset.

clusters
Cluster id to subset.

offset_amplitude
If TRUE, normalise CNA values for comparisons (z-score alike)

alpha
The parameter to select what to highlight.

Value

Examples

x = Rcongas::congas_example

# Default view
x %>% get_clusters_ploidy()

#> # A tibble: 140 x 8
#>    chr        from        to cluster      CN segment_mean lognorm_CN highlight
#>    <chr>     <dbl>     <dbl> <chr>     <dbl>        <dbl>      <dbl> <lgl>    
#>  1 chr1          1  12600000 1        0.0636         1.19       1.25 FALSE    
#>  2 chr1   12600001  60150000 1        0.0276         1.54       1.56 FALSE    
#>  3 chr1   61200001  77100000 1       -0.0279         1.28       1.25 FALSE    
#>  4 chr1   78750001 156000000 1        0.0287         1.56       1.59 FALSE    
#>  5 chr1  156000001 249300000 1        0.0183         1.46       1.48 FALSE    
#>  6 chr2          1  26250000 1        0.0899         1.59       1.68 FALSE    
#>  7 chr2   26250001  38100000 1        0.0373         1.28       1.32 FALSE    
#>  8 chr2   64050001  87150000 1        0.0266         1.46       1.49 FALSE    
#>  9 chr2   88350001 159900000 1       -0.0369         1.49       1.45 FALSE    
#> 10 chr2  159900001 176850000 1        0.0375         1.21       1.25 FALSE    
#> # … with 130 more rows

# Subset by chromosome
x %>% get_clusters_ploidy(chromosomes = 'chr1')

#> # A tibble: 10 x 8
#>    chr        from        to cluster      CN segment_mean lognorm_CN highlight
#>    <chr>     <dbl>     <dbl> <chr>     <dbl>        <dbl>      <dbl> <lgl>    
#>  1 chr1          1  12600000 1        0.0636         1.19       1.25 FALSE    
#>  2 chr1   12600001  60150000 1        0.0276         1.54       1.56 FALSE    
#>  3 chr1   61200001  77100000 1       -0.0279         1.28       1.25 FALSE    
#>  4 chr1   78750001 156000000 1        0.0287         1.56       1.59 FALSE    
#>  5 chr1  156000001 249300000 1        0.0183         1.46       1.48 FALSE    
#>  6 chr1          1  12600000 2       -0.0636         1.19       1.13 FALSE    
#>  7 chr1   12600001  60150000 2       -0.0276         1.54       1.51 FALSE    
#>  8 chr1   61200001  77100000 2        0.0279         1.28       1.31 FALSE    
#>  9 chr1   78750001 156000000 2       -0.0287         1.56       1.53 FALSE    
#> 10 chr1  156000001 249300000 2       -0.0183         1.46       1.44 FALSE    

# Subset by cluster id and chromosome
x %>% get_clusters_ploidy(chromosomes = 'chr1', clusters = "c1")

#> # A tibble: 0 x 8
#> # … with 8 variables: chr <chr>, from <dbl>, to <dbl>, cluster <chr>, CN <dbl>,
#> #   segment_mean <dbl>, lognorm_CN <dbl>, highlight <lgl>

# Change parameter to find what is most relevant
x %>% get_clusters_ploidy(alpha = 0.1)

#> # A tibble: 140 x 8
#>    chr        from        to cluster      CN segment_mean lognorm_CN highlight
#>    <chr>     <dbl>     <dbl> <chr>     <dbl>        <dbl>      <dbl> <lgl>    
#>  1 chr1          1  12600000 1        0.0636         1.19       1.25 FALSE    
#>  2 chr1   12600001  60150000 1        0.0276         1.54       1.56 FALSE    
#>  3 chr1   61200001  77100000 1       -0.0279         1.28       1.25 FALSE    
#>  4 chr1   78750001 156000000 1        0.0287         1.56       1.59 FALSE    
#>  5 chr1  156000001 249300000 1        0.0183         1.46       1.48 FALSE    
#>  6 chr2          1  26250000 1        0.0899         1.59       1.68 FALSE    
#>  7 chr2   26250001  38100000 1        0.0373         1.28       1.32 FALSE    
#>  8 chr2   64050001  87150000 1        0.0266         1.46       1.49 FALSE    
#>  9 chr2   88350001 159900000 1       -0.0369         1.49       1.45 FALSE    
#> 10 chr2  159900001 176850000 1        0.0375         1.21       1.25 FALSE    
#> # … with 130 more rows