Extract input data.

General data accessing getter function to return any of:

data,

segmentation,

normalisation factors.

The function uses the what parameter to return the appropriate type of information

Since the input formats are different, the outputs are also different based on the value of what, but they are always in tibble format.

Besides obvious columns, these information will also be available in the returned tibbles.

what = "data": modality and value_type, the latter reporting the likelihood code associated to the modality.

what = "segmentation":

ATAC_nonzerovals and ATAC_peaks, reporting the number of ATAC entries mapped to a segment, and the number of peaks these come from. Note that one non-zero entry is given by a cell that reported a ATAC count in a peak mapping within the segment.

RNA_nonzerovals and RNA_genes, reporting the number of RNA entries mapped to a segment, and the number of genes these come from. For RNA, one non-zero entry is given by a cell that reported an RNA count in a gene mapping within the segment.

what = "normalisation": just modality.

get_input(x, what = "data")

Arguments

x
An object of class rcongasplus.

what
Any of "data", "segmentation" or "normalisation".

Value

A tibble; its format depends on what. See the examples.

Examples

data(example_object)

# Extract the input data, after mapping to segments
get_input(example_object, what = 'data') 

#> # A tibble: 343,494 x 5
#>    segment_id               cell              value modality value_type
#>    <chr>                    <chr>             <dbl> <chr>    <chr>     
#>  1 chr1:143264541:193212282 AAACCCAAGTCTGGTT  32.0  RNA      G         
#>  2 chr1:143264541:193212282 AAACCCACAACTGCCG -10.7  RNA      G         
#>  3 chr1:143264541:193212282 AAACCCAGTAAGCGGT -72.9  RNA      G         
#>  4 chr1:143264541:193212282 AAACCCATCTCTTGCG  62.8  RNA      G         
#>  5 chr1:143264541:193212282 AAACGAAAGCTAAATG  19.1  RNA      G         
#>  6 chr1:143264541:193212282 AAACGAAGTATCTTCT  54.2  RNA      G         
#>  7 chr1:143264541:193212282 AAACGAAGTTTACCAG   9.37 RNA      G         
#>  8 chr1:143264541:193212282 AAACGAATCAAACGAA -11.6  RNA      G         
#>  9 chr1:143264541:193212282 AAACGAATCAAGTGGG   8.12 RNA      G         
#> 10 chr1:143264541:193212282 AAACGCTAGCATCAGG -25.5  RNA      G         
#> # … with 343,484 more rows

# e.g., in this way you can get input ATAC values
get_input(example_object, what = 'data') %>% dplyr::filter(modality == 'ATAC')

#> # A tibble: 179,766 x 5
#>    segment_id               cell                    value modality value_type
#>    <chr>                    <chr>                   <dbl> <chr>    <chr>     
#>  1 chr1:143264541:193212282 AAACGAACACGGCCAT-CRC_17    28 ATAC     NB        
#>  2 chr1:143264541:193212282 AAACGAACATGGAGGT-CRC_17    42 ATAC     NB        
#>  3 chr1:143264541:193212282 AAACGAAGTCAGTGTT-CRC_17    28 ATAC     NB        
#>  4 chr1:143264541:193212282 AAACGAAGTCCACCAG-CRC_17    20 ATAC     NB        
#>  5 chr1:143264541:193212282 AAACGAAGTGGACCAA-CRC_17    24 ATAC     NB        
#>  6 chr1:143264541:193212282 AAACGAATCACCACAA-CRC_17    50 ATAC     NB        
#>  7 chr1:143264541:193212282 AAACGAATCACGGGTC-CRC_17    29 ATAC     NB        
#>  8 chr1:143264541:193212282 AAACGAATCACGTGTA-CRC_17    17 ATAC     NB        
#>  9 chr1:143264541:193212282 AAACTCGAGCAAGGTC-CRC_17    54 ATAC     NB        
#> 10 chr1:143264541:193212282 AAACTCGAGTAATCAG-CRC_17    10 ATAC     NB        
#> # … with 179,756 more rows

 
# Extract the input segmentation. 
get_input(example_object, what = 'segmentation')

#> # A tibble: 36 x 9
#>    chr       from       to copies segment_id         ATAC_nonzerovals ATAC_peaks
#>    <chr>    <int>    <int>  <int> <chr>                         <dbl>      <dbl>
#>  1 chr1    1.69e4   1.22e8      3 chr1:16949:121596…            54609         11
#>  2 chr1    1.43e8   1.93e8      3 chr1:143264541:19…            19882          4
#>  3 chr1    2.14e8   2.32e8      4 chr1:214483397:23…             4998          1
#>  4 chr2    1.73e5   2.42e8      4 chr2:173150:24176…           114664         23
#>  5 chr3    3.20e5   9.02e7      3 chr3:319825:90229…            39621          8
#>  6 chr3    9.39e7   1.98e8      3 chr3:93874275:198…            44565          9
#>  7 chr4    3.73e5   4.89e7      3 chr4:372924:48892…            14762          3
#>  8 chr4    5.25e7   1.90e8      3 chr4:52466188:190…            63740         13
#>  9 chr5    1.69e5   4.59e7      4 chr5:169186:45891…            14847          3
#> 10 chr5    5.04e7   7.39e7      4 chr5:50402346:739…             4945          1
#> # … with 26 more rows, and 2 more variables: RNA_nonzerovals <dbl>,
#> #   RNA_genes <dbl>

 
# Extract the input normalisation factors
get_input(example_object, what = 'normalisation')

#> # A tibble: 9,547 x 3
#>    cell             normalisation_factor modality
#>    <chr>                           <dbl> <chr>   
#>  1 AAACCCAAGTCTGGTT                15619 RNA     
#>  2 AAACCCACAACTGCCG                28531 RNA     
#>  3 AAACCCAGTAAGCGGT                42057 RNA     
#>  4 AAACCCATCTCTTGCG                16651 RNA     
#>  5 AAACGAAAGCTAAATG                21235 RNA     
#>  6 AAACGAAGTATCTTCT                16750 RNA     
#>  7 AAACGAAGTTTACCAG                17389 RNA     
#>  8 AAACGAATCAAACGAA                24646 RNA     
#>  9 AAACGAATCAAGTGGG                21820 RNA     
#> 10 AAACGCTAGCATCAGG                29114 RNA     
#> # … with 9,537 more rows