Create a dataset.

This function creates a dataset (an object of class rcongasplus) by assembling multiple single-cell input measurements (ATAC and/or RNA data modalities), the input segmentation (from bulk DNA sequencing), and the per-cell normalisation factors for the data.

All input data are passed as tibbles; the input formats are as follows:

for single-cell ATAC/RNA data, the cell identifier, the genomic coordinates (chr, from, to) which refer either to an ATAC peak, or an RNA gene identifier, and a value reporting the reads mapped.

for the input segmentation, the genomic coordinates (chr, from, to) which refer to the segment, and the number of copies (i.e., DNA ploidy) of the segment.

for normalization factors the cell identifier, the actual normalisation_factor and the modality to wihch the factor refers to

This function receives also other parameters - e.g., the models likelihoods - which will determine the overall behaviour of the underlying model, and how data are preared for inference.

A Negative Binomial likelihood ("NB"), which works directly from raw counts data

A Gaussian likelihood ("G"), which requires a z-score transformation of the data. This consists in :

scaling raw counts by the input normalization factors;

computing z-scores per cell;

summing up z-scores per segment;

computing z-scores per segment;

center the z-scores mean to the input ploidy.

init(
  rna,
  atac,
  segmentation,
  normalisation_factors,
  rna_likelihood = "G",
  atac_likelihood = "NB",
  reference_genome = "GRCh38",
  description = "A (R)CONGAS+ model"
)

Arguments

rna
A tibble with single-cell RNA data.

atac
A tibble with single-cell ATAC data.

segmentation
A tibble with the input segmentation.

normalisation_factors
A tibble with the input per-cell normalisation factors.

rna_likelihood
Type of likelihood used for RNA data ("G" for Gaussian and ""NB for Negative Binomial). The RNA default is "G".

atac_likelihood
Type of likelihood used for ATAC data, with default "NB".

reference_genome
Either "GRCh38" or "hg19".

description
A model in-words description.

Value

An object of class rcongasplus

Examples

data("example_input")

# For instance, RNA data
example_input$x_rna %>% print

#> # A tibble: 438,133 x 6
#>    gene     cell       value chr        from        to
#>    <chr>    <chr>      <int> <chr>     <int>     <int>
#>  1 DPM1     SRR5578183     1 chr20  50934867  50958555
#>  2 SCYL3    SRR5578183    10 chr1  169849631 169894267
#>  3 C1orf112 SRR5578183    26 chr1  169662007 169854080
#>  4 CFH      SRR5578183     6 chr1  196652043 196747504
#>  5 STPG1    SRR5578183    73 chr1   24356999  24416934
#>  6 NIPAL3   SRR5578183     5 chr1   24415802  24472976
#>  7 CFTR     SRR5578183     1 chr7  117287120 117715971
#>  8 CYP51A1  SRR5578183     3 chr7   92112153  92134803
#>  9 KRIT1    SRR5578183     3 chr7   92198969  92246166
#> 10 RAD52    SRR5578183     1 chr12    911736    990053
#> # … with 438,123 more rows

# .. or ATAC data
example_input$x_atac %>% print

#> # A tibble: 27,303 x 6
#>    gene     cell             value chr        from        to
#>    <chr>    <chr>            <int> <chr>     <int>     <int>
#>  1 MRE11    SRR5578299 dfwfd    44 chr11  94415570  94493885
#>  2 HK1      SRR5578219 dfwfd    35 chr10  69269984  69401884
#>  3 MCPH1    SRR5578235 dfwfd    38 chr8    6406596   6648508
#>  4 ABCC6    SRR5578226 dfwfd    89 chr16  16148928  16223522
#>  5 ZNF213   SRR5578231 dfwfd    11 chr16   3129777   3142804
#>  6 CD84     SRR5578229 dfwfd     2 chr1  160541095 160579516
#>  7 KCTD10   SRR5578260 dfwfd    14 chr12 109448655 109477359
#>  8 RPS27AP1 SRR5578187 dfwfd    79 chr17  10258762  10259232
#>  9 UQCC3    SRR5578238 dfwfd    52 chr11  62670273  62673686
#> 10 KLHL21   SRR5578238 dfwfd    22 chr1    6590724   6614607
#> # … with 27,293 more rows

# .. and segmentation
example_input$x_segmentation %>% print

#> # A tibble: 22 x 4
#>    chr    from        to copies
#>    <chr> <int>     <int>  <int>
#>  1 chr1      0 249250621      2
#>  2 chr2      0 243199373      2
#>  3 chr3      0 198022430      2
#>  4 chr4      0 191154276      2
#>  5 chr5      0 180915260      2
#>  6 chr6      0 171115067      2
#>  7 chr7      0 159138663      2
#>  8 chr8      0 146364022      2
#>  9 chr9      0 141213431      2
#> 10 chr10     0 135534747      2
#> # … with 12 more rows

# .. and x_normalisation factors
example_input$x_normalisation_factors %>% print

#> # A tibble: 95 x 3
#>    cell       normalisation_factor modality
#>    <chr>                     <dbl> <chr>   
#>  1 SRR5578183                    1 RNA     
#>  2 SRR5578187                    1 RNA     
#>  3 SRR5578188                    1 RNA     
#>  4 SRR5578189                    1 RNA     
#>  5 SRR5578190                    1 RNA     
#>  6 SRR5578193                    1 RNA     
#>  7 SRR5578198                    1 RNA     
#>  8 SRR5578199                    1 RNA     
#>  9 SRR5578200                    1 RNA     
#> 10 SRR5578201                    1 RNA     
#> # … with 85 more rows

x = init(
  rna = example_input$x_rna,
  atac = example_input$x_atac,
  segmentation = example_input$x_segmentation,
  normalisation_factors = example_input$x_normalisation_factors,
  rna_likelihood = "G", 
  atac_likelihood = 'NB',
  description = 'My crazy model')

#>                         ┌────────────────────────────────┐
#>                         │                                │
#>                         │   (R)CONGAS+: My crazy model   │
#>                         │                                │
#>                         └────────────────────────────────┘

#> 

#> ── RNA modality (16.755992 Mb) 

#> 

#> → Input events: 438133

#> → Input cells: 53

#> → Input locations: 17200

#> → Entries mapped: 438133, with 0 outside input segments that will be discarded.

#> → Using likelihood: G.

#> ! Gaussian likelihood requires z-score representation of input values.

#> → Normalising RNA counts using input normalisation factors.

#> → Computing z-score.

#> ! There are 2994 z-scores that are NA, will be removed.

#> # A tibble: 2,994 x 10
#>    gene  cell  value chr     from     to segment_id modality value_mean value_sd
#>    <chr> <chr> <dbl> <chr>  <int>  <int> <chr>      <chr>         <dbl>    <dbl>
#>  1 TBX15 SRR5…   NaN chr1  1.19e8 1.19e8 chr1:0:24… RNA               1        0
#>  2 PLAU  SRR5…    NA chr10 7.39e7 7.39e7 chr10:0:1… RNA               5       NA
#>  3 TRIM… SRR5…    NA chr11 5.59e7 5.59e7 chr11:0:1… RNA               2       NA
#>  4 EDN2  SRR5…    NA chr1  4.15e7 4.15e7 chr1:0:24… RNA               1       NA
#>  5 CHL1  SRR5…    NA chr3  1.97e5 4.09e5 chr3:0:19… RNA               1       NA
#>  6 DNAJ… SRR5…    NA chr2  2.19e8 2.19e8 chr2:0:24… RNA              51       NA
#>  7 RAPG… SRR5…   NaN chr7  2.21e7 2.24e7 chr7:0:15… RNA               1        0
#>  8 HMCN1 SRR5…    NA chr1  1.86e8 1.86e8 chr1:0:24… RNA               1       NA
#>  9 ACTA1 SRR5…   NaN chr1  2.29e8 2.29e8 chr1:0:24… RNA               1        0
#> 10 MUC4  SRR5…    NA chr3  1.96e8 1.96e8 chr3:0:19… RNA               1       NA
#> # … with 2,984 more rows

#> ! Center the new scores around the ploidy values.

#> 

#> ── ATAC modality (1.62176 Mb) 

#> 

#> → Input events: 27303

#> → Input cells: 42

#> → Input locations: 11192

#> → Entries mapped: 27303, with 0 outside input segments that will be discarded.

#> → Using likelihood: NB.

#> 

#> ── Checking segmentation 

#> 

#> ✓ Nothing to process.

  
print(x)

#> ── [ (R)CONGAS+ ] My crazy model ───────────────────────────────────────────────

#> 

#> ── CNA segments (reference: GRCh38) 

#> → Input 22 CNA segments, mean ploidy 2.

#> 
#> 	 | | | | | | | | | | | | | | | | | | | | | 
#> 
#> 	 Ploidy:    0     1     2     3     4     *   

#> 

#> ── Modalities 

#> → RNA: 53 cells with 17200 mapped genes, 438133 non-zero values. Likelihood: Gaussian.

#> → ATAC: 42 cells with 11192 mapped peaks, 27303 non-zero values. Likelihood: Negative Binomial.

#> 

#> ! Clusters: not available.

#> [1] 0