Compute the statistics of a dataset.

The statistics of a dataset regard the input data, and the fits. These are accessed by using different values for the what parameter passed to this function (by default, data are used).

The result value is a named list with several information that are usefull to summarise the model. The names should be self explicatory.

For data the following information are reported:

it is reported the number of cells in each modality,

the number of non-zero ATAC/RNA input entries used to create the object

the number of RNA genes and ATAC peaks used

the number of segments

the number of modalities

the keyword for each modality ("RNA" or "ATAC")

the likelihood for each modality ("G" for Gaussian or "NB" for Negative Binomial)

For fits the following information are reported:

...

If fits are not available, NULL is returned.

stat(x, what = "data")

Arguments

x
An object of class rcongasplus.

what
The type of statistics that needs to be computed. It can be one of "data" or "fit". By deafult it is "data".

Value

A named list with the available statistics.

Examples

data(example_object)

# Compare the information here 
print(example_object)

#> ── [ (R)CONGAS+ ] SU008 TUMOR 30 segments (73.66% genome) ──────────────────────

#> 

#> ── CNA segments (reference: GRCh38) 

#> → Input 30 CNA segments, mean ploidy 3.2.

#> 
#> 	 | | |  | | | |  | | | |  | | |  | |  |   |  |  | 
#> 
#> 	 Ploidy:    0     1     2     3     4     5     *   

#> 

#> ── Modalities 

#> → RNA: 714 cells with 8613 mapped genes, 1401728 non-zero values. Likelihood: Negative Binomial.

#> → ATAC: 259 cells with 284316 mapped peaks, 3083691 non-zero values. Likelihood: Negative Binomial.

#> ! Clusters: not available.

#> 

#> ──  LOG  ──

#> 

#> - 2021-03-30 17:58:41 Created input object.
#> - 2021-03-30 17:58:43 Filtered outliers: [6|0.05|0.95]

#> [1] 0

# ... to the one reported here for data
stat(example_object, what = 'data')

#> $ncells_RNA
#> [1] 714
#> 
#> $ncells_ATAC
#> [1] 259
#> 
#> $rna_events
#> [1] 1401728
#> 
#> $rna_genes
#> [1] 8613
#> 
#> $atac_events
#> [1] 3083691
#> 
#> $atac_peaks
#> [1] 284316
#> 
#> $nsegments
#> [1] 30
#> 
#> $nmodalities
#> [1] 2
#> 
#> $modalities
#> [1] "RNA"  "ATAC"
#> 
#> $rna_dtype
#> [1] "NB"
#> 
#> $atac_dtype
#> [1] "NB"
#> 
#> $zero_counts_cells_RNA
#> NULL
#> 
#> $zero_counts_cells_ATAC
#> NULL
#> 

# ... and here for fits
stat(example_object, what = 'fit')

#> NULL